[Triton] NFC: Split Triton ABI refinement into a separate pass
