+++ compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/Transforms/BUILD.bazel
@@ -4,7 +4,7 @@
 # See https://llvm.org/LICENSE.txt for license information.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-load("//build_tools/bazel:build_defs.oss.bzl", "iree_gentbl_cc_library", "iree_tablegen_doc", "iree_td_library")
+load("//build_tools/bazel:build_defs.oss.bzl", "iree_gentbl_cc_library")
 
 package(
     default_visibility = ["//visibility:public"],
@@ -49,6 +49,7 @@ cc_library(
         "ConvertTritonToFlowDispatch.cpp",
         "OutlineTritonCalls.cpp",
         "Passes.cpp",
+        "RefineTritonAbi.cpp",
     ],
     hdrs = ["Passes.h"],
     deps = [
